requirements.txt
django==1.11
future
six
pubchempy
numpy
scipy
bioservices
seaborn
plotly
celery
django-bulk_update
django_celery_results
django-filter==1.1
django_tables2
django_bootstrap3
django_tables2_column_shifter
django_sb_admin
django-rest-framework
django-gfiles
tablib
bioblend
django-autocomplete-light
